requirements.txt
seleniumbase>=4.47
rembg[cpu]>=2.0.70
PyQt5>=5.15
Pillow>=10.0
requests>=2.31
numpy>=1.24
scipy>=1.10
